requirements.txt
flask==2.0.1
python-dotenv==0.19.0
openai==0.27.0
functions-framework==3.*
firebase-admin==6.0.1
gunicorn==20.1.0
python-docx==0.8.11 